// src/lib/study-lenses--deprecated-architecture/orchestrate/lib/editing/types.ts
type LintDiagnostic = {
	readonly line: number;
	readonly column: number;
	readonly endLine?: number;
	readonly endColumn?: number;
	readonly severity: 'error' | 'warning' | 'rejection';
	readonly message: string;
	readonly source?: string;
	readonly entry?: DocEntry;
};
// ...
type DocEntry = {
	readonly description: string;
	readonly isJEJ: boolean;
	readonly example?: string;
	readonly commonMistakes?: readonly string[];
	readonly whenToUse?: string;
	readonly whyNotInJej?: string;
};